+++ addon/locales/ru/translations.js
@@ -161,7 +161,10 @@ export default {
       'setting': 'Настройка ',
       'is-saved': ' сохранена',
       'col-width-on': 'Включить настройку ширины столбцов',
-      'per-page': 'Записей на странице'
+      'per-page': 'Записей на странице',
+      'unresizable': 'Фиксированная ширина',
+      'unresizable-row-toolbar': 'Фиксированная ширина столбца инструментов',
+      'unresizable-row-menu': 'Фиксированная ширина столбца меню'
     },
 
     'form-load-time-tracker': {