+++ tests/dummy/app/routes/ember-flexberry-dummy-suggestion-edit.js
@@ -35,7 +35,8 @@ export default EditFormRoute.extend(EditFormRouteOperationsIndicationMixin, {
     suggestionUserVotesGroupEdit: {
       'DEFAULT': {
         'columnWidths': [{ 'propName': 'voteType', 'width': 133 }, { 'propName': 'author', 'width': 348 }, { 'propName': 'author.eMail', 'width': 531 }],
-        'sorting': [{ 'propName': 'author', 'direction': 'asc' }]
+        'sorting': [{ 'propName': 'author', 'direction': 'asc' }],
+        'unresizableColumns': ['OlvRowToolbar']
       }
     },
     filesGroupEdit: {
@@ -47,7 +48,8 @@ export default EditFormRoute.extend(EditFormRouteOperationsIndicationMixin, {
     },
     suggestionCommentsGroupEdit: {
       'DEFAULT': {
-        'sorting': [{ 'propName': 'votes', 'direction': 'asc' }, { 'propName': 'moderated', 'direction': 'desc' }, { 'propName': 'text', 'direction': 'asc' }]
+        'sorting': [{ 'propName': 'votes', 'direction': 'asc' }, { 'propName': 'moderated', 'direction': 'desc' }, { 'propName': 'text', 'direction': 'asc' }],
+        'unresizableColumns': ['votes']
       }
     }
   },